requirements.txt
flask>=2.0.0
werkzeug==2.2.3
pillow>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-image>=0.18.0
qrcode>=7.0
tqdm>=4.60.0
pandas>=1.3.0
psutil>=5.8.0

# Optional dependencies for extended explainability
transformers>=4.18.0  # For Vision Transformer support
torch-geometric>=2.0.0  # For Graph Neural Network support
networkx>=2.6.0  # For graph visualization 